Fix warning

+++ Test/ve2/Source/ReqLogin.cpp
@@ -97,7 +97,7 @@ std::cout << "Got session ID but we don't know the object ID for the local playe
         Strings strs = Split(locStr, ',');
         if (strs.size() == 4)
         {
-          Vec3f pos(ToInt(strs[0]), ToInt(strs[1]), ToInt(strs[2]));
+          Vec3f pos(ToFloat(strs[0]), ToFloat(strs[1]), ToFloat(strs[2]));
           TheGSWaitForNewLocation::Instance()->SetStartPos(pos);
           loc = ToInt(strs[3]);
         }
